Implement googletest
Two build types have been defined. A HOST build and a non HOST build. A
HOST build uses the host's compilers (generally amd64) and is mainly
used for unittesting. Non host-build is for ARM targets and the ARM
compiler is used in said case.

The googletest suite is therefore used for host builds and runs unit
tests. googletest has been added as a submodule and the cmake build
setup to use it on host builds.