Remove file, function and code line  from logs
In such a small program it is kind of obvious where messages are coming
from, and any context coming from knowing the originating file and
function does not give that much benefit, therefore the extra info has
been removed.

On the other hand this removal makes the logs that much cleaner and
easier to read.