Add temperature graph on display
The graph takes up most of the screen and shows a history of the
temperature changes throughout time. It is automatically updated
whenever `graphicsSetTemperature` is called.

The temperature label is still shown on screen but it has been moved, it
is now shown on the bottom left corner of the screen, its size has also
been diminished.