Add basic SSR control
Basic functionality to control a SSR based on AC cycles has been added. Upon
clicking the start button, the SSR GPIO is turned on/off based on the desired duty
cycle. The duty cycle represents the amount of AC cycles to turn on the SSR. Said
cycles are counted based on the square wave output of a zero cross detector. 

This functionality uses GPIO IRQs, they are all set and run on cortex 0. This is
important as gpio_set_irq_enabled enables IRQs only on the core the 
function was called from.