feat: Implement HTTP client
An HTTP client capable of creating reflow sessions on the API server and
logging measurements has been added. It uses lwip's API to write the
payloads directly into the TCP sockets.

Temperature measurements are sent in groups of 10 readings, every 10
seconds. While the software is not meant to recover from failure,
failures related to tx/rx won't stop the rest of the application from
working.