feat: support stop button
Reflow process can be stopped on a falling edge on GPIO 22. This falling
edge is meant to come from a push button. Care must be taken to run
control tasks on a single core since they all access the stage variable,
this works as a crude but effective sync method.